Fix typo in test suite for describe('defaults')
Currently, the test suite says, 'should copy a property if that key is
already set on the target.' However, default should only copy a property
if the property is NOT already set on the target.

Furthermore, the next test says 'should not copy a proprety if that key is
already set on the target', so these two tests say contradictory statements.